Reconcile how activation works in RS1 and RS3.
In RS1, the holographic space must be created during SetWindow, otherwise the app goes flat.

On RS3 we can delay creation of the holographic space until OnActivated. We want to delay creating the holographic space until we know the activation context such that we create the proper view: holographic or flat.